refactor(db): Adjust student_id type in student_courses table
Changed the data type of the student_id column in the student_courses table from BIGINT to INT.
This adjustment aligns the schema with the expected range of student identifiers and may offer minor storage optimizations.

+++ src/main/resources/db/migration/V7__create_student_courses_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE student_courses (
-    student_id BIGINT NOT NULL,
+    student_id INT NOT NULL,
     course VARCHAR(255),
     grade VARCHAR(255),
     PRIMARY KEY (student_id, course),
